requirements.txt
streamlit
sentence-transformers
google-generativeai
PyPDF2
python-docx
openpyxl
pandas
faiss-cpu
langchain
